backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
feedparser==6.0.10
wbgapi==2.0.0
pandas==2.1.3
pydantic==2.5.0
